requirements.txt
# FastAPI and ASGI server
fastapi==0.104.1
uvicorn[standard]==0.24.0

# Authentication and security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6

# Configuration and environment
pydantic[email]==2.5.0
pydantic-settings==2.0.3
python-dotenv==1.0.0

# Rate limiting and middleware
slowapi==0.1.9

# Testing
pytest==7.4.3
pytest-asyncio==0.21.1
httpx==0.25.2

# Development tools (optional)
black==23.11.0
isort==5.12.0
flake8==6.1.0
mypy==1.7.1

# Additional utilities
email-validator==2.1.0